+++ src/components/Header/RecentView.tsx
@@ -1,30 +1,16 @@
 import styles from "./RecentView.module.scss";
 import classNames from "classnames/bind";
 import { Typography } from "../index";
-import { useEffect, useState } from "react";
-import IndexedDBSingleton, {
-  RECENT_VIEW,
-  RecentViewedPokemonType,
-} from "utils/IndexedDB/IndexedDBSingleton";
-import { useLocation, useNavigate } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import useIndexChange from "hooks/useIndexChange";
+import { useRecentPokemonList } from "store/recentPokemon";
 
 const cx = classNames.bind(styles);
 export default function RecentView() {
   const navigate = useNavigate();
-  const { pathname } = useLocation();
 
-  const [recentPokemonList, setRecentPokemonList] = useState<
-    RecentViewedPokemonType[]
-  >([]);
-
+  const { recentPokemonList } = useRecentPokemonList();
   const { index } = useIndexChange(recentPokemonList.length);
-
-  useEffect(() => {
-    IndexedDBSingleton.getAllRecentPokemon(RECENT_VIEW).then((res) => {
-      setRecentPokemonList(res);
-    });
-  }, [pathname]);
 
   return (
     <>
